% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{as_code}
\alias{as_code}
\title{\code{as_code()} converts a vector into R code that generates it.}
\usage{
as_code(x)
}
\arguments{
\item{x}{vector}
}
\description{
\code{as_code()} converts a vector into R code that generates it.
}
