% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{breaks_log10}
\alias{breaks_log10}
\alias{labels_log10}
\title{Axis breaks: scale_*_log10(breaks=breaks_log10)}
\usage{
breaks_log10(limits)

labels_log10(breaks)
}
\arguments{
\item{limits}{numeric}

\item{breaks}{numeric}
}
\description{
Axis breaks: scale_*_log10(breaks=breaks_log10)

Axis label formatter: scale_*_log10(labels=labels_log10)
}
