% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioc.R
\name{library_bioc}
\alias{library_bioc}
\title{Shortcut for \code{source()} and \code{biocLite()}}
\usage{
library_bioc(mirror = getOption("BioC_mirror",
  "https://bioconductor.org"), ...)
}
\arguments{
\item{mirror}{URL for BioC_mirror}

\item{...}{passed to \code{biocLite()}}
}
\description{
Shortcut for \code{source()} and \code{biocLite()}
}
