% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{ord}
\alias{ord}
\alias{chr}
\title{Conversion between character and integer code point}
\usage{
ord(char)

chr(i)
}
\arguments{
\item{char}{character}

\item{i}{integer}
}
\description{
\code{ord()} is a pythonic alias of \code{charToRaw()}.

\code{chr()} is a pythonic alias of \code{intToUtf8()}.
}
