% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column.R
\name{dedfcol}
\alias{dedfcol}
\alias{dedfcol_all}
\alias{demtrxcol}
\alias{demtrxcol_all}
\title{Handle columns with data.frame and matrix}
\usage{
dedfcol(.data, at)

dedfcol_all(.data)

demtrxcol(.data, at)

demtrxcol_all(.data)
}
\arguments{
\item{.data}{A data.frame.}

\item{at}{target column.}
}
\description{
\code{dedfcol()} and \code{dedfcol_all()} dissolve data.frame columns in a tibble.

\code{demtrxcol()} and \code{demtrxcol_all()} dissolve matrix columns in a tibble.
}
