% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figprocess.R
\name{oxipng}
\alias{oxipng}
\alias{cwebp}
\title{Functions for knitr fig.process}
\usage{
oxipng(input, options = NULL)

cwebp(input, options = NULL)
}
\arguments{
\item{input}{A path to an image file.}

\item{options}{A list.}
}
\description{
\code{oxipng()} losslessly improves compression of PNG files.

\code{cwebp()} compresses a raster image to WebP format.
}
