% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstats.R
\name{fst_HBK}
\alias{fst_HBK}
\alias{fst_HSM}
\title{Calculate Kst by Hudson, Boos, and Kaplan (1992)}
\usage{
fst_HBK(within, between, n = 2)

fst_HSM(within, between)
}
\arguments{
\item{within, between}{mean branch length or diversity}

\item{n}{number of subpopulations}
}
\value{
numeric

numeric
}
\description{
Calculate Kst by Hudson, Boos, and Kaplan (1992)

Calculate Fst by Hudson, Slatkin, and Maddison (1992)
}
