% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggaes.R
\name{aes_try}
\alias{aes_try}
\alias{aes_try_}
\alias{aes_try_string}
\title{Variants of \code{ggplot2::aes()} to ignore errors in evaluation}
\usage{
aes_try(.data, ...)

aes_try_(.data, ...)

aes_try_string(.data, ...)
}
\arguments{
\item{.data}{data.frame}

\item{...}{passed to \code{ggplot2::aes()}}
}
\description{
Variants of \code{ggplot2::aes()} to ignore errors in evaluation
}
