% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglabel.R
\name{breaks_log10}
\alias{breaks_log10}
\alias{labels_log10}
\title{Helper functions to make labels in ggplot}
\usage{
breaks_log10(limits)

labels_log10(breaks)
}
\arguments{
\item{limits}{numeric}

\item{breaks}{numeric}
}
\description{
Helper functions to make labels in ggplot
}
\details{
\code{breaks_log10} makes powers of ten for
the \code{breaks} argument of \code{ggplot2::scale_*_log10()}.

\code{labels_log10} is an axis label formatter to be passed to
the \code{labels} argument of \code{ggplot2::scale_*_log10()}.
}
\examples{
ggplot2::ggplot(tibble::tibble(x = seq_len(10), y = 2 ** x)) +
  ggplot2::geom_point(aes(x = x, y = y)) +
  scale_y_log10(breaks = breaks_log10, labels = labels_log10)
}
