% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmisc.R
\name{insert_layer}
\alias{insert_layer}
\alias{mean_sd}
\alias{label_none}
\title{Misc functions with ggplot2}
\usage{
insert_layer(p, ..., after = 0L)

mean_sd(x, mult = 1)

label_none(labels, ...)
}
\arguments{
\item{p}{gg object}

\item{...}{layers to add}

\item{after}{position; 0L is back}

\item{x}{numeric vector.}

\item{mult}{number of multiples of standard error.}

\item{labels}{Data frame of labels. Usually contains only one
element, but faceting over multiple factors entails multiple
label variables.}
}
\description{
\code{insert_layer()} insert layers to an arbitrary position

\code{mean_sd()} calculates mean and sd for \code{\link[ggplot2:stat_summary]{ggplot2::stat_summary()}}.

\code{label_none()} removes \code{strip.text} while keeping \code{strip.background}.
}
\examples{
(ggplot2::ggplot(ggplot2::mpg) +
  ggplot2::aes(drv, hwy) +
  ggplot2::geom_boxplot(fill = "gold") +
  ggplot2::geom_dotplot(binaxis = "y", stackdir = "center", binwidth = 1, stackratio = 0.5) +
  ggplot2::stat_summary(fun.data = mean_sd, color = "red", size = 2, linewidth = 2) +
  ggplot2::facet_wrap(ggplot2::vars(drv), scale = "free_x", label = label_none)) |>
  insert_layer(ggplot2::geom_violin(fill = "darkorange"), after = -2L)
}
