% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggridges.R
\name{ridges2bars}
\alias{ridges2bars}
\title{Helper functions to enhance ggridges package}
\usage{
ridges2bars(.data, x, height, width = 0.9)
}
\arguments{
\item{.data}{data.frame}

\item{x, height}{Columns to transform}

\item{width}{Bar width.}
}
\description{
Helper functions to enhance ggridges package
}
\details{
\code{\link[=ridges2bars]{ridges2bars()}} adds data points for \code{\link[ggridges:geom_ridgeline]{ggridges::geom_ridgeline()}}
to make bar-shaped ridgelines.
}
\examples{
library(magrittr)
df_bars = tidyr::crossing(lambda = c(1, 2, 3), x = seq(0, 8)) \%>\%
  dplyr::mutate(height = dpois(x, lambda)) \%>\%
  ridges2bars(x, height) \%>\%
  print()
ggplot2::ggplot(df_bars) +
  ggplot2::aes(x, lambda) +
  ggridges::geom_ridgeline(ggplot2::aes(height = height, group = lambda))
ggplot2::ggplot(df_bars) +
  ggplot2::aes(lambda, x) +
  ggridges::geom_vridgeline(ggplot2::aes(width = height, group = lambda))
}
