% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{viewport_layout}
\alias{grid_grob}
\alias{print.grob}
\alias{viewport_layout}
\title{Shortcut for grid::viewport()}
\usage{
viewport_layout(row, col)

grid_grob(grobs, nrow = NULL, ncol = NULL, byrow = FALSE)

\method{print}{grob}(x, ...)
}
\arguments{
\item{row}{layout.pos.row}

\item{col}{layout.pos.col}

\item{grobs}{list of grobs}

\item{nrow, ncol}{integer}

\item{byrow}{logical}

\item{x}{An object of class \code{"grob"} or NULL.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
grid::viewport

grid::gTree
}
\description{
Shortcut for grid::viewport()

Just another gridExtra::arrangeGrob()

Shortcut for grid::grid.draw()
}

