% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph.R
\name{igraph_layout}
\alias{igraph_layout}
\title{Utilities for igraph class}
\usage{
igraph_layout(graph, layout = igraph::nicely(), ...)
}
\arguments{
\item{graph}{The input graph.}

\item{layout}{The layout specification. It must be a call
to a layout specification function.}

\item{...}{Further modifiers, see a complete list below.
For the \code{print} methods, it is ignored.}
}
\value{
tibble
}
\description{
Utilities for igraph class
}
