% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{crossing_rep}
\alias{crossing_rep}
\title{Shortcut for tidyr::crossing() with repeats.}
\usage{
crossing_rep(x, times = 1L)
}
\arguments{
\item{x}{atomic vector}

\item{times}{the number of repeats}
}
\value{
tbl
}
\description{
Shortcut for tidyr::crossing() with repeats.
}
