% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr-stan.R
\name{knit_engines_set_cache_stan}
\alias{knit_engines_set_cache_stan}
\title{Register knitr engine for stan code}
\usage{
knit_engines_set_cache_stan(cache_stan_prefix = NULL)
}
\arguments{
\item{cache_stan_prefix}{Prefix for output files.
A relative path is evaluated from the source file.}
}
\description{
This is a simple alternative to \code{\link[cmdstanr:register_knitr_engine]{cmdstanr::register_knitr_engine()}}.
Chunk code is not compiled, but just written to a file with decent caching
to facilitate showing the usage of \code{\link[cmdstanr:cmdstan_model]{cmdstanr::cmdstan_model()}}.
}
\details{
Stan file is automatically named after the chunk label, .i.e.,
\code{"{cache_stan_prefix}{label}.stan"}.

Set \code{cache_stan} to \code{TRUE} or \code{FALSE} for stan chunks so that
\code{output.var} can be omitted. Then set it to the chunk label of stan code
for dependent R chunks.
}
