% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr-time.R
\name{knit_hooks_set_timeit}
\alias{knit_hooks_set_timeit}
\title{Register knitr hook to time chunks}
\usage{
knit_hooks_set_timeit(enable = TRUE)
}
\arguments{
\item{enable}{A logical.}
}
\description{
Register knitr hook to time chunks
}
