% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{setdefault}
\alias{setdefault}
\alias{update_}
\alias{select_list}
\title{Utilities for lists}
\usage{
setdefault(object, other = NULL, ...)

update_(object, other = NULL, ...)

select_list(object, ...)
}
\arguments{
\item{object, other}{A list.}

\item{...}{Name-value pairs or tidyselect expressions.}
}
\description{
\code{setdefault()} and \code{update_()} return lists modified in similar ways to
the equivalent methods of Python dict.

\code{select_list()} is an alternative to \code{\link[dplyr:select]{dplyr::select()}} for lists.
}
