% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo.R
\name{plot_logo}
\alias{plot_logo}
\alias{save_logo}
\alias{save_sticker}
\title{Draw and save the logo}
\usage{
plot_logo(
  devsize = grDevices::dev.size(),
  expand = 0,
  swoosh = "#e08010",
  dot = "#a4321a",
  path = "#222222"
)

save_logo(
  filename,
  height = 4,
  dpi = 300,
  ...,
  expand = 0,
  bg = NULL,
  midground = NULL,
  width = height,
  swoosh = "#e08010",
  dot = "#a4321a",
  path = "#222222"
)

save_sticker(filename, height = 4, dpi = 300, ...)
}
\arguments{
\item{devsize}{A numeric vector. \code{c(width, height)}}

\item{expand}{A numeric.}

\item{swoosh, dot, path}{Color}

\item{filename}{File name to create on disk.}

\item{dpi}{Plot resolution. Also accepts a string input: "retina" (320),
"print" (300), or "screen" (72). Applies only to raster output types.}

\item{...}{Other arguments passed on to the graphics device function,
as specified by \code{device}.}

\item{bg}{Background colour. If \code{NULL}, uses the \code{plot.background} fill value
from the plot theme.}

\item{midground}{Layers to insert behind the logo.}

\item{width, height}{Size in inches.}
}
\description{
See \code{tools/logo.R} for usage.
}
