% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{permutate}
\alias{permutate}
\alias{probit}
\alias{logit}
\alias{sigmoid}
\alias{logistic}
\alias{exponent}
\alias{mantissa}
\alias{scientific_notation}
\title{Shortcut to mathematical functions}
\usage{
permutate(n, k)

probit(p)

logit(p, base = exp(1))

sigmoid(x, gain = 1)

logistic(x, r = 1, k = 1, x0 = 0)

exponent(x, base = exp(1))

mantissa(x, base = exp(1))

scientific_notation(x, base = 10)
}
\arguments{
\item{n, k}{numeric vectors}

\item{p}{a numeric vector between 0 and 1}

\item{base}{a numeric}

\item{x}{a numeric vector}

\item{gain}{a numeric}

\item{r}{a numeric vector}

\item{x0}{a numeric vector}
}
\description{
Shortcut to mathematical functions

\code{scientific_notation} splits number into mantissa and exponent
}
