% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moment.R
\name{moment}
\alias{moment}
\alias{skewness}
\alias{kurtosis}
\alias{bimodality}
\title{Moment for arbitrary order}
\usage{
moment(x, order, n = length(x))

skewness(x)

kurtosis(x)

bimodality(x)
}
\arguments{
\item{x}{numeric vector}

\item{order}{integer}

\item{n}{length of x}
}
\description{
Moment for arbitrary order

kurtosis - 3 is called excess kurtosis, which is 0 for normal distribution
}
