% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{chk}
\alias{chk}
\title{Package development utilities}
\usage{
chk(pkg = ".", install = FALSE, vignettes = FALSE, args = "--timings", ...)
}
\arguments{
\item{pkg}{The package to use, can be a file path to the package or a
package object.  See \code{\link[devtools:as.package]{as.package()}} for more information.}

\item{install}{If FALSE, add \code{--no-install} to \code{args}}

\item{vignettes}{If FALSE, add \code{--ignore-vignettes} to \code{args}}

\item{args}{An optional character vector of additional command line
arguments to be passed to \verb{R CMD INSTALL}. This defaults to the
value of the option \code{"devtools.install.args"}.}

\item{...}{additional arguments passed to \code{\link[remotes:install_deps]{remotes::install_deps()}}
when installing dependencies.}
}
\description{
\code{chk()} is a thin wrapper of \code{\link[devtools:check]{devtools::check()}}.
}
