% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{clean_install}
\alias{clean_install}
\title{Package utilities}
\usage{
clean_install(pkg = ".", upgrade = FALSE, ...)
}
\arguments{
\item{pkg}{package description, can be path or package name.  See
\code{\link[=as.package]{as.package()}} for more information}

\item{upgrade}{One of "ask", "always" or "never". "ask" prompts the user for
which out of date packages to upgrade. For non-interactive sessions "ask" is
equivalent to "always". \code{TRUE} and \code{FALSE} are also accepted and
correspond to "always" and "never" respectively.}

\item{...}{additional arguments passed to \code{\link[remotes:install_deps]{remotes::install_deps()}}
when installing dependencies.}
}
\description{
Package utilities
}
\details{
\code{clean_install} is a shortcut of \code{devtools::clean_dll} and
\code{devtools::install}.
}
