% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paf.R
\name{as_paf}
\alias{as_paf}
\alias{read_paf}
\alias{write_paf}
\alias{read_axt}
\alias{read_chain}
\alias{read_net}
\alias{read_singmaf}
\title{Unify pairwise alignment files to Pairwise mApping Format (PAF)}
\usage{
as_paf(x)

read_paf(file)

write_paf(x, file)

read_axt(file)

read_chain(file)

read_net(file)

read_singmaf(file)
}
\arguments{
\item{x}{data.frame.}

\item{file}{paths to alignment files.}
}
\description{
PAF is described in \href{https://lh3.github.io/minimap2/minimap2.html}{minimap2 manual}.
}
\seealso{
\url{https://genome.ucsc.edu/goldenPath/help/axt.html}

\url{https://genome.ucsc.edu/goldenPath/help/chain.html}

\url{https://genome.ucsc.edu/goldenPath/help/net.html}

\url{https://genome.ucsc.edu/FAQ/FAQformat.html#format5}
}
