% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{isdir}
\alias{isdir}
\alias{splitext}
\title{A shortcut}
\usage{
isdir(...)

splitext(path)
}
\arguments{
\item{...}{character vectors conatining file paths}

\item{path}{a string}
}
\value{
a logical vector

a character vector (root, ext)
}
\description{
A shortcut

os.path.splitext() in Python
}

