% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo.R
\name{node_labels}
\alias{node_labels}
\alias{named_edges}
\alias{phylo2igraph}
\title{Utilities for phylo class of ape package}
\usage{
node_labels(x)

named_edges(x)

phylo2igraph(x)
}
\arguments{
\item{x}{phylo object}
}
\description{
Utilities for phylo class of ape package
}
