% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{max_print}
\alias{max_print}
\alias{pipeshell}
\alias{\%|\%}
\alias{less}
\alias{grepp}
\alias{\%G\%}
\alias{redirect}
\alias{adjust_print_options}
\title{Utilities for printing and piping}
\usage{
max_print(x, max.print = getOption("max.print"), width = Inf, ...)

pipeshell(x, command, ...)

x \%|\% command

less(x, ...)

grepp(x, expr, ...)

x \%G\% expr

redirect(expr, file = "/dev/null")

adjust_print_options(max.print = 30L)
}
\arguments{
\item{x}{an object to print}

\item{max.print}{maximum number of rows to print}

\item{width}{maximum number of columns to print}

\item{...}{further arguments passed to \code{print}}

\item{command}{to which the text is sent; string or expression.}

\item{expr}{arguments to \code{grep -P}; string or expression.}

\item{file}{a writable \link{connection} or a character string naming the
    file to write to, or \code{NULL} to stop sink-ing.}
}
\description{
Utilities for printing and piping
}
\details{
\code{max_print} prints as many elements in a big tibble as possible.

\code{pipeshell} and \code{\%|\%} send \code{x} to \code{command} via shell.

\code{less} print \code{x} in \code{getOption('pager')}. The operation is in-memory and
efficient than \code{page(x, method='print')} that involves a temporary file.

\code{grepp} and \code{\%G\%} are shorthand for \code{x \%|\% "grep -P {expr}"}.

\code{redirect} evaluates \code{expr} while \code{sink(file)} is in operation.

\code{adjust_print_options} sets width and height according to the current environment.
}
\examples{
\dontrun{
max_print(mpg)

mpg \%>\% less()
mpg \%>\% grepp(manual)
mpg \%>\% grepp("manual")
mpg \%>\% pipeshell(wc)

iris \%|\% wc
iris \%|\% "wc -l"
iris \%|\% `grep seto | less`

vir = "seto"
iris \%G\% vir
iris \%G\% !!vir
iris \%G\% `-v seto | less`
}
}
