% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{randint}
\alias{randint}
\title{Generate random integers between [a, b]}
\usage{
randint(n, a, b)
}
\arguments{
\item{n}{integer}

\item{a}{integer}

\item{b}{integer}
}
\value{
n integer vector
}
\description{
Generate random integers between [a, b]
}
\examples{
table(randint(600L, 1L, 6L))
}

