% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{scan_matrix}
\alias{scan_matrix}
\alias{read_boost_ini}
\alias{read_cb}
\title{Functions to read files}
\usage{
scan_matrix(filename, what = double(0), sep = "", skip = 0L,
  header = FALSE)

read_boost_ini(file)

read_cb(...)
}
\arguments{
\item{filename}{a string}

\item{what}{type of data}

\item{sep}{a string}

\item{skip}{an integer}

\item{header}{a logical}

\item{file}{filename or text}

\item{...}{passed to \code{read.table()}}
}
\description{
\code{scan_matrix} scans a simple CSV and make a matrix.

\code{read_boost_ini} reads an INI-like config file of boost::program_options.

\code{read_cb} reads paste board into data.frame.
}
