% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{rsplit}
\alias{rsplit}
\title{Pythonic method to manupilate strings}
\usage{
rsplit(string, pattern, n = 42)
}
\arguments{
\item{string}{a character vector}

\item{pattern}{a separator string as ICU regex}

\item{n}{the number of splits to do}
}
\value{
a list of character vectors
}
\description{
Pythonic method to manupilate strings
}

