% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symlink.R
\name{symlink}
\alias{symlink}
\alias{link_resolve}
\title{Utility functions to handle symlinks}
\usage{
symlink(path, target, follow = TRUE)

link_resolve(path)
}
\arguments{
\item{path}{A character vector.}

\item{target}{A character vector. If an existing directory is given, links are created there.}

\item{follow}{A logical vector. \code{FALSE} means \code{ln -h} on macOS and \code{ln -n} on Linux.}
}
\description{
\code{symlink()} is a simple wrapper of \code{\link[fs:create]{fs::link_create()}}.
Existing links are replaced if link paths are different.

\code{link_resolve()} is a non-stop variant of \code{\link[fs:path_math]{fs::path_real()}}.
}
