% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex.R
\name{powerintf}
\alias{powerintf}
\alias{write_tabular}
\title{Print power of ten}
\usage{
powerintf(fmt, n)

write_tabular(x, file = "", col.names = TRUE)
}
\arguments{
\item{fmt}{printf string}

\item{n}{numeric}

\item{x}{data.frame}

\item{file}{string}

\item{col.names}{logical}
}
\value{
string
}
\description{
Print power of ten

Translate data.frame to LaTeX table
}
