% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_clean}
\alias{axis_line}
\alias{theme_clean}
\alias{theme_wtl}
\title{Incomplete theme that does not overwrite other elements}
\usage{
theme_clean(base_size = 12, base_family = "sans")

theme_wtl(base_size = 12, base_family = "sans")

axis_line(colour = NULL, size = NULL, linetype = NULL, lineend = NULL)
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}

\item{colour}{line colour}

\item{size}{line size}

\item{linetype}{line type}

\item{lineend}{line end}
}
\description{
Incomplete theme that does not overwrite other elements

Complete theme based on theme_bw()

Set L-shaped axes
}

