% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{today}
\alias{today}
\alias{now}
\title{Snippet for time in R}
\usage{
today(format = "\%Y\%m\%d")

now(format = "\%Y\%m\%d_\%H\%M\%S")
}
\arguments{
\item{format}{string}
}
\description{
Snippet for time in R
}
