% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toml.R
\name{toTOML}
\alias{toTOML}
\alias{format_iso8601}
\title{Simple converter from list to TOML.}
\usage{
toTOML(data, usetz = TRUE)

format_iso8601(x, usetz = TRUE, precision = NULL, ...)
}
\arguments{
\item{data}{A list.}

\item{usetz}{Include the time zone in the formatting.  If
\code{usetz} is \code{TRUE}, the time zone is included. If
\code{usetz} is \code{"Z"}, the time is converted to "UTC" and
the time zone is indicated with "Z" ISO8601 notation.}

\item{x}{An object to convert to ISO8601 character format.}

\item{precision}{The amount of precision to represent with
substrings of "ymdhms", as "y"ear, "m"onth, "d"ay, "h"our,
"m"inute, and "s"econd. (e.g. "ymdhm" would show precision
through minutes.  When \code{NULL}, full precision for the object
is shown.}

\item{...}{Additional arguments to methods.}
}
\description{
\code{format_iso8601()} is compliant with TOML.
}
