% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{easierprof}
\alias{easierprof}
\title{Simple wrapper of Rprof() and summaryRprof()}
\usage{
easierprof(expr, interval = 0.02, memory = FALSE)
}
\arguments{
\item{expr}{R expression}

\item{interval}{numeric}

\item{memory}{logical}
}
\value{
summaryRprof()
}
\description{
Simple wrapper of Rprof() and summaryRprof()
}

