% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtl-package.R
\docType{package}
\name{wtl-package}
\alias{wtl}
\alias{wtl-package}
\title{wtl: Personal R package}
\description{
Personal R package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/heavywatal/rwtl}
  \item Report bugs at \url{https://github.com/heavywatal/rwtl/issues}
}

}
\author{
\strong{Maintainer}: Watal M. Iwasaki \email{heavywatal@gmail.com} (\href{https://orcid.org/0000-0002-4718-7605}{ORCID})

}
\keyword{internal}
