% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioc.R
\name{library_bioc}
\alias{library_bioc}
\title{source('*/biocLite.R') and biocLite()}
\usage{
library_bioc(mirror = getOption("BioC_mirror"), ask = FALSE)
}
\arguments{
\item{mirror}{URL for BioC_mirror}

\item{ask}{passed to biocLite()}
}
\description{
source('*/biocLite.R') and biocLite()
}
