% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{bioc_install}
\alias{bioc_install}
\title{Install or update Bioconductor, CRAN, and GitHub packages}
\usage{
bioc_install(pkgs, ..., update = FALSE, ask = interactive())
}
\arguments{
\item{pkgs}{\code{character()} vector of package names to install or
update.  A missing value updates installed packages according
to \verb{update =} and \verb{ask =}. Package names containing a '/' are
treated as GitHub repositories and installed using
\code{remotes::install_github()}.}

\item{...}{Additional arguments used by \code{install.packages()}.}

\item{update}{\code{logical(1)}. When \code{FALSE}, \code{BiocManager::install()}
does not attempt to update old packages. When \code{TRUE}, update
old packages according to \code{ask}.}

\item{ask}{\code{logical(1)} indicating whether to prompt user before
installed packages are updated.  If TRUE, user can choose
whether to update all outdated packages without further
prompting, to pick packages to update, or to cancel updating
(in a non-interactive session, no packages will be updated
unless \code{ask = FALSE}).}
}
\value{
\code{BiocManager::install()} returns the \code{pkgs} argument, invisibly.
}
\description{
\code{bioc_install()} is a thin wrapper of \code{\link[BiocManager:install]{BiocManager::install()}}.
}
