% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{bioc_install}
\alias{bioc_install}
\alias{bioc_valid}
\title{Install or update Bioconductor, CRAN, and GitHub packages}
\usage{
bioc_install(
  ...,
  update = FALSE,
  ask = interactive(),
  type = binary_if_macos()
)

bioc_valid(..., type = binary_if_macos())
}
\arguments{
\item{...}{Additional arguments used by \code{install.packages()}.}

\item{update}{\code{logical(1)}. When \code{FALSE}, \code{BiocManager::install()}
does not attempt to update old packages. When \code{TRUE}, update
old packages according to \code{ask}.}

\item{ask}{\code{logical(1)} indicating whether to prompt user before
installed packages are updated.  If TRUE, user can choose
whether to update all outdated packages without further
prompting, to pick packages to update, or to cancel updating
(in a non-interactive session, no packages will be updated
unless \code{ask = FALSE}).}

\item{type}{character(1) The type of available package (e.g.,
binary, source) to check validity against; see
\verb{\\link{available.packages}()}.}
}
\value{
\code{BiocManager::install()} returns the \code{pkgs} argument, invisibly.
}
\description{
Install or update Bioconductor, CRAN, and GitHub packages
}
\details{
\code{bioc_install()} and \code{bioc_valid()} are thin wrappers of BiocManager functions.
}
