% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/command.R
\name{command_args}
\alias{command_args}
\alias{this_file}
\title{get commandline arguments}
\usage{
command_args()

this_file()
}
\value{
a list

a character; NULL in interactive mode
}
\description{
get commandline arguments

script path in Rscript
}

