% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/command.R
\name{command_args}
\alias{command_args}
\title{Parse command line arguments}
\usage{
command_args()
}
\description{
Parse command line arguments
}
\details{
\code{command_args} is an extention of \code{commandArgs()}.
}
