% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{center_range}
\alias{center_range}
\alias{class_at}
\alias{filter_duplicated}
\title{Utilities for data.frame}
\usage{
center_range(.data, .cols)

class_at(.data, .cols)

filter_duplicated(.data, ...)
}
\arguments{
\item{.data}{A data.frame.}

\item{.cols}{tidy-select passed to \code{\link[dplyr:across]{dplyr::across()}}}

\item{...}{tidy-select passed to \code{\link[dplyr:select]{dplyr::select()}}}
}
\description{
\code{center_range()} mutates the selected columns for zero-centered plot.

\code{class_at()} summarizes the class of the selected columns.

\code{filter_duplicated()} extracts duplicated rows including first occurrences.
}
