% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{move_left_}
\alias{move_left}
\alias{move_left_}
\alias{mutate_left}
\alias{mutate_left_}
\title{Move specified columns to left}
\usage{
move_left_(.data, ..., .dots)

move_left(.data, ...)

mutate_left_(.data, .dots)

mutate_left(.data, ...)
}
\arguments{
\item{.data}{tbl}

\item{...}{colnames or expressions}

\item{.dots}{vector of strings or expressions}
}
\description{
Move specified columns to left
}

