% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{mutate_left}
\alias{mutate_left}
\alias{dedfcol}
\alias{dedfcol_all}
\alias{demtrxcol}
\alias{demtrxcol_all}
\alias{center_range}
\alias{class_at}
\title{Utilities for data.frame}
\usage{
mutate_left(.data, ...)

dedfcol(.data, at)

dedfcol_all(.data)

demtrxcol(.data, at)

demtrxcol_all(.data)

center_range(.data, .cols)

class_at(.data, .cols)
}
\arguments{
\item{.data}{a data.frame.}

\item{...}{Additional arguments for the function calls in \code{.fns}. Using these
\code{...} is strongly discouraged because of issues of timing of evaluation.}

\item{at}{target column.}

\item{.cols}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Columns to transform.
Because \code{across()} is used within functions like \code{summarise()} and
\code{mutate()}, you can't select or compute upon grouping variables.}
}
\description{
\code{mutate_left()} adds new columns to the left.

\code{dedfcol()} and \code{dedfcol_all()} dissolve data.frame columns in a tibble.

\code{demtrxcol()} and \code{demtrxcol_all()} dissolve matrix columns in a tibble.
}
