% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{move_left}
\alias{move_left}
\alias{mutate_left}
\alias{dedfcol}
\alias{dedfcol_all}
\alias{center_range}
\alias{class_at}
\title{Utilities for data.frame}
\usage{
move_left(.data, ...)

mutate_left(.data, ...)

dedfcol(.data, at)

dedfcol_all(.data)

center_range(.data, ...)

class_at(.data, ...)
}
\arguments{
\item{.data}{tibble}

\item{...}{colnames or expressions}

\item{at}{target column}
}
\description{
Utilities for data.frame
}
\details{
\code{move_left} moves specified columns to the left.

\code{mutate_left} adds new columns to the left.

\code{dedfcol} and \code{dedfcol_all} dissolve data.frame columns in a tibble.
}
