% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{move_left}
\alias{move_left}
\alias{mutate_left}
\alias{center_range}
\alias{class_at}
\title{Move specified columns to left}
\usage{
move_left(.data, ...)

mutate_left(.data, ...)

center_range(.data, ...)

class_at(.data, ...)
}
\arguments{
\item{.data}{tibble}

\item{...}{colnames or expressions}
}
\value{
tbl
}
\description{
Move specified columns to left
}
