% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{center_range}
\alias{center_range}
\alias{class_at}
\title{Utilities for data.frame}
\usage{
center_range(.data, .cols)

class_at(.data, .cols)
}
\arguments{
\item{.data}{A data.frame.}

\item{.cols}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Columns to transform.
You can't select grouping columns because they are already automatically
handled by the verb (i.e. \code{\link[dplyr:summarise]{summarise()}} or \code{\link[dplyr:mutate]{mutate()}}).}
}
\description{
Utilities for data.frame
}
