% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{move_left}
\alias{move_left}
\alias{mutate_left}
\alias{mutate_jitter}
\alias{center_range}
\alias{class_at}
\title{Utilities for data.frame}
\usage{
move_left(.data, ...)

mutate_left(.data, ...)

mutate_jitter(.data, ..., amount)

center_range(.data, ...)

class_at(.data, ...)
}
\arguments{
\item{.data}{tibble}

\item{...}{colnames or expressions}

\item{amount}{magnitude of random variations to add}
}
\description{
\code{move_left} moves specified columns to the left.

\code{mutate_left} adds new columns to the left.

\code{mutate_jitter} adds random variations to given columns.
}
