% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{shannon_index}
\alias{evenness}
\alias{shannon_index}
\alias{simpson_index}
\title{Shannon index}
\usage{
shannon_index(freqs, base = exp(1))

simpson_index(freqs)

evenness(species)
}
\arguments{
\item{freqs}{a numeric vector}

\item{base}{a numeric}

\item{species}{a factor vector}
}
\value{
a numeric

a numeric

a number
}
\description{
Shannon index

Simpson index

H / H_max
}

