% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{shannon_index}
\alias{shannon_index}
\alias{simpson_index}
\alias{evenness}
\title{Diversity index}
\usage{
shannon_index(freqs, base = exp(1))

simpson_index(freqs)

evenness(species)
}
\arguments{
\item{freqs}{a numeric vector}

\item{base}{a positive or complex number: the base with respect to which
    logarithms are computed.  Defaults to \eqn{e}=\code{exp(1)}.}

\item{species}{a factor vector}
}
\description{
\code{evenness()} calculates H / H_max
}
