% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exec.R
\name{exec_knit}
\alias{exec_knit}
\alias{exec_render}
\title{Shortcut to use knitr and rmarkdown from commandline}
\usage{
exec_knit(input, output = NA)

exec_render(input, output_dir = NA)
}
\arguments{
\item{input}{File path.}

\item{output}{File path.}

\item{output_dir}{Directory path.}
}
\description{
Shortcut to use knitr and rmarkdown from commandline
}
