% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs.R
\name{is_outdated}
\alias{is_outdated}
\title{Functions related to file systems}
\usage{
is_outdated(path, prerequisites = fs::path())
}
\arguments{
\item{path}{files to test}

\item{prerequisites}{comparables}
}
\description{
Functions related to file systems
}
\details{
\code{is_outdated()} tests existence, size, and modification time of files.
Note that it returns \code{TRUE} if a file is empty.
}
