% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs.R
\name{is_outdated}
\alias{is_outdated}
\alias{file_copy_try}
\alias{path_destination}
\title{Functions related to file systems}
\usage{
is_outdated(path, prerequisites = character(0L))

file_copy_try(path, new_path, overwrite = FALSE)

path_destination(path, new_path)
}
\arguments{
\item{path}{files to test}

\item{prerequisites}{comparables}

\item{new_path}{path to the new file or target directory.}

\item{overwrite}{If this is FALSE and the file is outdated, a message will be shown.}
}
\description{
\code{is_outdated()} tests existence and modification time of files.

\code{file_copy_try()} is a modified version of \code{\link[fs:copy]{fs::file_copy()}} to copy only new files.

\code{path_destination()} normalizes destination paths.
}
