% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggannotate.R
\name{annotate_regpolygon}
\alias{annotate_regpolygon}
\title{Variations of \code{ggplot2::annotate()}}
\usage{
annotate_regpolygon(
  n,
  radius = 1,
  x = 0,
  y = 0,
  start = pi/2,
  linejoin = "mitre",
  ...
)
}
\arguments{
\item{n}{The number of sides.}

\item{radius}{The distance from the center to a vertex.}

\item{x, y}{The center coordinates.}

\item{start}{The angle of the starting point.}

\item{linejoin, ...}{Extra arguments passed to \code{ggplot2::annotate("polygon")}.}
}
\description{
\code{annotate_regpolygon()} is a shortcut of \code{ggplot2::annotate("polygon")}
to draw a regular polygon.
}
\examples{
ggplot2::ggplot() +
  annotate_regpolygon(6L, radius = 5, linewidth = 6, color = "#333333", fill = "#ffffff") +
  annotate_regpolygon(6L, radius = 2, fill = "#C41A41", linetype = 0, linewidth = 0) +
  ggplot2::coord_fixed()
}
