% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggannotate.R
\name{annotate_polygon}
\alias{annotate_polygon}
\title{Variations of \code{ggplot2::annotate()}}
\usage{
annotate_polygon(x, y, radius = 1, stroke = 1, n = 6L, ...)
}
\arguments{
\item{x, y}{center coordinate}

\item{radius}{size of polygon}

\item{stroke}{line width}

\item{n}{number of sides}

\item{...}{passed to \code{ggplot2::annotate()}}
}
\description{
\code{annotate_polygon()} is a shortcut of \code{ggplot2::annotate("polygon")}
}
