% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggannotate.R
\name{annotate_polygon}
\alias{annotate_polygon}
\alias{annotate_polygon2}
\title{Variations of \code{ggplot2::annotate()}}
\usage{
annotate_polygon(x, y, n = 6L, radius = 1, stroke = 1, ...)

annotate_polygon2(x, y, n = 6L, radius = 1, stroke = 1,
  colour = "#666666", linejoin = "mitre", ...)
}
\arguments{
\item{x, y}{center coordinate}

\item{n}{number of sides}

\item{radius}{size of polygon}

\item{stroke}{line width}

\item{...}{passed to \code{ggplot2::annotate("polygon")}}

\item{colour, linejoin}{passed to \code{ggplot2::annotate("path")}}
}
\description{
Variations of \code{ggplot2::annotate()}
}
\details{
\code{annotate_polygon()} is a shortcut of \code{ggplot2::annotate("polygon")}.

\code{annotate_polygon2()} has more control on line stroke.
}
\examples{
ggplot2::ggplot() +
  annotate_polygon2(0, 0, n = 6L, radius = 5, fill = "#ffffff", colour = "#333333", stroke = 6) +
  annotate_polygon(0, 0, n = 6L, radius = 2, fill = "#C41A41")
}
