% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmisc.R
\name{gghist}
\alias{gghist}
\alias{ggsave_quartz}
\title{Misc functions with ggplot2}
\usage{
gghist(x, ...)

ggsave_quartz(filename, plot = ggplot2::last_plot(),
  device = grDevices::quartz, path = NULL, scale = 1, width = 7,
  height = 7, units = c("in", "cm", "mm"), dpi = 300, limitsize = TRUE)
}
\arguments{
\item{x}{vector}

\item{...}{passed to geom_*()}

\item{filename}{File name to create on disk.}

\item{plot}{Plot to save, defaults to last plot displayed.}

\item{device}{Device to use. Can be either be a device function
(e.g. \code{\link[=png]{png()}}), or one of "eps", "ps", "tex" (pictex),
"pdf", "jpeg", "tiff", "png", "bmp", "svg" or "wmf" (windows only).}

\item{path}{Path to save plot to (combined with filename).}

\item{scale}{Multiplicative scaling factor.}

\item{width}{Plot size in \code{units} ("in", "cm", or "mm").
If not supplied, uses the size of current graphics device.}

\item{height}{Plot size in \code{units} ("in", "cm", or "mm").
If not supplied, uses the size of current graphics device.}

\item{units}{Plot size in \code{units} ("in", "cm", or "mm").
If not supplied, uses the size of current graphics device.}

\item{dpi}{Plot resolution. Also accepts a string input: "retina" (320),
"print" (300), or "screen" (72). Applies only to raster output types.}

\item{limitsize}{When \code{TRUE} (the default), \code{ggsave} will not
save images larger than 50x50 inches, to prevent the common error of
specifying dimensions in pixels.}
}
\description{
\code{gghist} is a simple alternative to \code{hist()}

\code{ggsave_quartz} is a shortcut of \code{ggplot2::ggsave()} with quartz device
}
