% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtheme.R
\name{theme_wtl}
\alias{theme_wtl}
\alias{erase}
\alias{axis_line}
\title{Shortcut for ggplot theme}
\usage{
theme_wtl(base_size = 12, base_family = "")

erase(..., .names = NULL)

axis_line(
  colour = NULL,
  size = NULL,
  linetype = NULL,
  lineend = NULL,
  color = NULL,
  arrow = NULL,
  inherit.blank = FALSE
)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}

\item{...}{elements to be blank}

\item{.names}{string vector of element names}

\item{colour, color}{Line/border colour. Color is an alias for colour.}

\item{size}{text size in pts.}

\item{linetype}{Line type. An integer (0:8), a name (blank, solid,
dashed, dotted, dotdash, longdash, twodash), or a string with
an even number (up to eight) of hexadecimal digits which give the
lengths in consecutive positions in the string.}

\item{lineend}{Line end Line end style (round, butt, square)}

\item{arrow}{Arrow specification, as created by \code{\link[grid:arrow]{grid::arrow()}}}

\item{inherit.blank}{Should this element inherit the existence of an
\code{element_blank} among its parents? If \code{TRUE} the existence of
a blank element among its parents will cause this element to be blank as
well. If \code{FALSE} any blank parent element will be ignored when
calculating final element state.}
}
\description{
\code{theme_wtl()} is a slightly modified \link[ggplot2:ggtheme]{ggplot2::theme_bw}.

\code{erase()} is a shortcut for \code{theme(... = element_blank())}.

\code{axis_line()} sets L-shaped axes.
}
