% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{pch_plot}
\alias{col2hex}
\alias{pch_plot}
\title{Plot pch variations}
\usage{
pch_plot()

col2hex(color = grDevices::colors())
}
\arguments{
\item{color}{a character vector}
}
\value{
a character vector
}
\description{
Plot pch variations

Translate coloname to hex
}

