% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{pch_plot}
\alias{pch_plot}
\alias{gghist}
\alias{col2hex}
\title{Utilities for graphics}
\usage{
pch_plot(alpha = 1, stroke = 1)

gghist(x, ..., binwidth = NULL, bins = NULL)

col2hex(color = grDevices::colors())
}
\arguments{
\item{alpha, stroke}{passed to \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}}

\item{x}{vector}

\item{...}{passed to \code{\link[ggplot2:geom_histogram]{ggplot2::geom_histogram()}} or \code{\link[ggplot2:geom_bar]{ggplot2::geom_bar()}}}

\item{binwidth}{The width of the bins. Can be specified as a numeric value
or as a function that takes x after scale transformation as input and
returns a single numeric value. When specifying a function along with a
grouping structure, the function will be called once per group.
The default is to use the number of bins in \code{bins},
covering the range of the data. You should always override
this value, exploring multiple widths to find the best to illustrate the
stories in your data.

The bin width of a date variable is the number of days in each time; the
bin width of a time variable is the number of seconds.}

\item{bins}{Number of bins. Overridden by \code{binwidth}. Defaults to 30.}

\item{color}{a character vector}
}
\description{
\code{pch_plot()} shows pch variations.

\code{gghist()} is a simple alternative to \code{\link[=hist]{hist()}}

\code{col2hex()} translates color name to hex.
}
