% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{pch_plot}
\alias{axis_line}
\alias{col2hex}
\alias{ggsave_quartz}
\alias{pch_plot}
\alias{theme_clean}
\alias{theme_wtl}
\title{Plot pch variations}
\usage{
pch_plot()

col2hex(color = grDevices::colors())

theme_clean(base_size = 12, base_family = "sans")

theme_wtl(base_size = 12, base_family = "sans")

axis_line(colour = NULL, size = NULL, linetype = NULL, lineend = NULL)

ggsave_quartz(filename, plot = ggplot2::last_plot(),
  device = grDevices::quartz, path = NULL, scale = 1, width = 7,
  height = 7, units = c("in", "cm", "mm"), dpi = 300, limitsize = TRUE)
}
\arguments{
\item{color}{a character vector}

\item{base_size}{base font size}

\item{base_family}{base font family}

\item{colour}{line colour}

\item{size}{line size}

\item{linetype}{line type}

\item{lineend}{line end}

\item{filename}{File name to create on disk.}

\item{plot}{Plot to save, defaults to last plot displayed.}

\item{device}{Device to use (function or any of the recognized extensions,
e.g. \code{"pdf"}). By default, extracted from filename extension.
\code{ggsave} currently recognises eps/ps, tex (pictex), pdf, jpeg, tiff,
png, bmp, svg and wmf (windows only).}

\item{path}{Path to save plot to (combined with filename).}

\item{scale}{Multiplicative scaling factor.}

\item{width}{Plot dimensions, defaults to size of current graphics
device.}

\item{height}{see ggplot2::ggsave}

\item{units}{Units for width and height when specified explicitly (in, cm,
or mm)}

\item{dpi}{Resolution used for raster outputs.}

\item{limitsize}{When \code{TRUE} (the default), \code{ggsave} will not
save images larger than 50x50 inches, to prevent the common error of
specifying dimensions in pixels.}
}
\value{
a character vector
}
\description{
Plot pch variations

Translate coloname to hex

Incomplete theme that does not overwrite other elements

Complete theme

Set L-shaped axes

Save PDF with quartz
}

