% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmisc.R, R/graphics.R
\name{mean_sd}
\alias{mean_sd}
\alias{pch_plot}
\alias{gghist}
\alias{col2hex}
\title{Utilities for graphics}
\usage{
mean_sd(x, mult = 1.96)

pch_plot()

gghist(x, ...)

col2hex(color = grDevices::colors())
}
\arguments{
\item{x}{vector}

\item{mult}{number of multiples of standard error}

\item{...}{passed to geom_*()}

\item{color}{a character vector}
}
\description{
\code{mean_sd} calculates mean and sd for \code{ggplot2::stat_summary()}.

\code{pch_plot} shows pch variations.

\code{gghist} is a simple alternative to \code{hist()}

\code{col2hex} translates color name to hex.
}
