% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{pch_plot}
\alias{pch_plot}
\alias{col2hex}
\alias{gghist}
\alias{ggsave_quartz}
\title{Plot pch variations}
\usage{
pch_plot()

col2hex(color = grDevices::colors())

gghist(x, ...)

ggsave_quartz(filename, plot = ggplot2::last_plot(),
  device = grDevices::quartz, path = NULL, scale = 1, width = 7,
  height = 7, units = c("in", "cm", "mm"), dpi = 300, limitsize = TRUE)
}
\arguments{
\item{color}{a character vector}

\item{x}{vector}

\item{...}{passed to geom_*()}

\item{filename}{File name to create on disk.}

\item{plot}{Plot to save, defaults to last plot displayed.}

\item{device}{Device to use. Can be either be a device function
(e.g. \code{\link{png}}), or one of "eps", "ps", "tex" (pictex),
"pdf", "jpeg", "tiff", "png", "bmp", "svg" or "wmf" (windows only).}

\item{path}{Path to save plot to (combined with filename).}

\item{scale}{Multiplicative scaling factor.}

\item{width}{Plot size in \code{units} ("in", "cm", or "mm").
If not supplied, uses the size of current graphics device.}

\item{height, units}{see ggplot2::ggsave}

\item{dpi}{Plot resolution. Applies only to raster output types.}

\item{limitsize}{When \code{TRUE} (the default), \code{ggsave} will not
save images larger than 50x50 inches, to prevent the common error of
specifying dimensions in pixels.}
}
\value{
a character vector

gg
}
\description{
Plot pch variations

Translate coloname to hex

Simple alternative to hist() with ggplot2

Save PDF with quartz
}
