% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{viewport_layout}
\alias{as_gg}
\alias{grid_grob}
\alias{print.grob}
\alias{read_png}
\alias{viewport_layout}
\title{Shortcut for grid::viewport()}
\usage{
viewport_layout(row, col)

grid_grob(grobs, nrow = NULL, ncol = NULL, byrow = FALSE)

\method{print}{grob}(x, ...)

read_png(file)

as_gg(grob)
}
\arguments{
\item{row}{layout.pos.row}

\item{col}{layout.pos.col}

\item{grobs}{list of grobs}

\item{nrow, ncol}{integer}

\item{byrow}{logical}

\item{x}{An object of class \code{"grob"} or NULL.}

\item{...}{further arguments passed to or from other methods.}

\item{file}{path to PNG file}

\item{grob}{graphical object}
}
\value{
grid::viewport

grid::gTree

grob

ggplot
}
\description{
Shortcut for grid::viewport()

Just another gridExtra::arrangeGrob()

Shortcut for grid::grid.draw()

Read PNG file as grob

Convert grob to gg
}

