% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{viewport_layout}
\alias{as_gg}
\alias{gprint}
\alias{grid_grob}
\alias{png_grob}
\alias{viewport_layout}
\title{Shortcut for grid::viewport()}
\usage{
viewport_layout(row, col)

grid_grob(grobs, nrow = NULL, ncol = NULL, byrow = FALSE)

gprint(x, newpage = TRUE, recording = FALSE)

png_grob(file)

as_gg(grob)
}
\arguments{
\item{row}{layout.pos.row}

\item{col}{layout.pos.col}

\item{grobs}{list of grobs}

\item{nrow, ncol}{integer}

\item{byrow}{logical}

\item{x}{An object of class \code{"grob"} or NULL.}

\item{newpage}{logical}

\item{recording}{A logical value to indicate whether the drawing
    operation should be recorded on the Grid display list.}

\item{file}{path to PNG file}

\item{grob}{graphical object}
}
\value{
grid::viewport

grid::gTree

grob

ggplot
}
\description{
Shortcut for grid::viewport()

Just another gridExtra::arrangeGrob()

Shortcut for grid::grid.draw()

Read PNG file as grob

Convert grob to gg
}

