% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{gprint}
\alias{gprint}
\alias{png_grob}
\alias{as_gg}
\title{Shortcut for grid::grid.draw()}
\usage{
gprint(x, newpage = TRUE, recording = FALSE)

png_grob(file)

as_gg(grob, base_size = 12, base_family = "sans")
}
\arguments{
\item{x}{An object of class \code{"grob"} or NULL.}

\item{newpage}{logical}

\item{recording}{A logical value to indicate whether the drawing
    operation should be recorded on the Grid display list.}

\item{file}{path to PNG file}

\item{grob}{graphical object}

\item{base_size, base_family}{see ggplot2::ggtheme}
}
\value{
grob

ggplot
}
\description{
Shortcut for grid::grid.draw()

Read PNG file as grob

Convert grob to gg
}
