% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{crossing_rep}
\alias{crossing_rep}
\title{Shortcut for tidyr::crossing() with repeats.}
\usage{
crossing_rep(x, times = 1L)
}
\arguments{
\item{x}{a vector (of any mode including a \code{\link[base]{list}}) or a factor or (for
    \code{rep} only) a \code{POSIXct} or \code{POSIXlt} or \code{Date}
    object; or an S4 object containing such an object.}
}
\value{
tbl
}
\description{
Shortcut for tidyr::crossing() with repeats.
}
