% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo.R
\name{logo_plot}
\alias{logo_plot}
\alias{logo_save}
\alias{logo_save_sticker}
\alias{logo_svg_optimize}
\title{Draw and save the logo}
\usage{
logo_plot(
  devsize = grDevices::dev.size(),
  expand = 0,
  swoosh = "#E08010",
  dot = "#A4321A",
  path = "#202020"
)

logo_save(
  filename,
  height = 1.5,
  dpi = 300,
  ...,
  expand = 0,
  bg = NULL,
  midground = NULL,
  width = height,
  swoosh = "#E08010",
  dot = "#A4321A",
  path = "#202020"
)

logo_save_sticker(filename, height = 4, dpi = 300, ...)

logo_svg_optimize(filename, path = "#F0F0F0")
}
\arguments{
\item{devsize}{A numeric vector. \code{c(width, height)}}

\item{expand}{A numeric.}

\item{swoosh, dot, path}{Color}

\item{filename}{File name to create on disk.}

\item{dpi}{Plot resolution. Also accepts a string input: "retina" (320),
"print" (300), or "screen" (72). Applies only to raster output types.}

\item{...}{Other arguments passed on to the graphics device function,
as specified by \code{device}.}

\item{bg}{Background colour. If \code{NULL}, uses the \code{plot.background} fill value
from the plot theme.}

\item{midground}{Layers to insert behind the logo.}

\item{width, height}{Size in inches.}
}
\description{
See \code{tools/logo.R} for usage.
}
