% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{permutate}
\alias{permutate}
\alias{probit}
\alias{logit}
\alias{sigmoid}
\alias{logistic}
\alias{heaviside}
\alias{exponent}
\alias{mantissa}
\alias{scientific_notation}
\title{Permutation}
\usage{
permutate(n, k)

probit(p)

logit(p, base = exp(1))

sigmoid(x, gain = 1)

logistic(x, r = 1, k = 1, x0 = 0)

heaviside(x, h0 = NA)

exponent(x, base = exp(1))

mantissa(x, base = exp(1))

scientific_notation(x, base = 10)
}
\arguments{
\item{n, }{k numeric vectors}

\item{k}{a numeric vector}

\item{p}{a numeric vector [0, 1]}

\item{base}{a numeric}

\item{x}{a numeric vector}

\item{gain}{a numeric}

\item{r}{a numeric vector}

\item{x0}{a numeric vector}

\item{h0}{return value if x==0}
}
\value{
numeric

a numeric vector

a numeric vector

a numeric vector

a numeric vector

a numeric vector
}
\description{
Permutation

Probit function

Logit functon: log-odds; the inverse of sigmoid

Sigmoid function: the inverse of logit

Logistic function

Heaviside step function

exponent

mantissa, significand, coefficient

Split number into mantissa and exponent
}
