% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms.R
\name{ms}
\alias{ms}
\alias{parse_ms}
\title{Execute ms as an external program}
\usage{
ms(nsam = 4L, nreps = 2L, theta = 5)

parse_ms(lines)
}
\arguments{
\item{nsam}{number of samples}

\item{nreps}{number of repeats}

\item{theta}{population mutation rate}

\item{lines}{output of system('ms')}
}
\value{
string vector

list of string vector
}
\description{
Execute ms as an external program

Read ms output and return a list of '01' string vector
}
