% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{lib_upgrade}
\alias{lib_upgrade}
\alias{old_packages}
\alias{install_packages}
\title{Package management utilities}
\usage{
lib_upgrade(..., binary = TRUE, ask = interactive())

old_packages(..., binary = TRUE, bioc = FALSE)

install_packages(pkg, lib = .libPaths()[[1L]], ..., binary = TRUE)
}
\arguments{
\item{...}{Passed to \code{\link[utils:update.packages]{utils::old.packages()}} or \code{\link[pak:pkg_install]{pak::pkg_install()}}.}

\item{binary}{A logical. Set \code{FALSE} to accept source packages.}

\item{ask}{Whether to ask for confirmation when installing a different
version of a package that is already installed. Installations that only
add new packages never require confirmation.}

\item{bioc}{A logical. Set \code{TRUE} to include BioC repositories.}

\item{pkg}{Package names or remote package specifications to install.
See \link[pak:pak_package_sources]{pak package sources} for details.}

\item{lib}{Package library to install the packages to. Note that \emph{all}
dependent packages will the be installed here, even if they are
already installed in another library.}
}
\description{
\code{lib_upgrade()} is a stopgap until \code{pak::lib_upgrade()} \url{https://github.com/r-lib/pak/issues/168}.
Set \code{PKG_PLATFORMS} before loading pak: \url{https://github.com/r-lib/pak/issues/412}.
There is no direct way to enforce binary for now: \url{https://github.com/r-lib/pak/issues/318}.
\code{pak::pkg_upgrade()} is not available yet: \url{https://github.com/r-lib/pak/pull/289}.
Versioned CRAN packages are not implemented yet \url{https://github.com/r-lib/pkgdepends/blob/main/R/type-cran.R}.

\code{old_packages()} is a thin wrapper of \code{\link[utils:update.packages]{utils::old.packages()}}.

\code{install_packages()} is a thin wrapper of \code{\link[pak:pkg_install]{pak::pkg_install()}}
that can temporarily disable forced binary installation with \code{PKG_PLATFORMS}.
}
