% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{brewer_palette}
\alias{brewer_palette}
\alias{scale_colour_gradientb}
\alias{scale_fill_gradientb}
\title{Wrapper of brewer.pal()}
\usage{
brewer_palette(name, n = 0L)

scale_colour_gradientb(palette, n = 0L, reverse = FALSE, ...)

scale_fill_gradientb(palette, n = 0L, reverse = FALSE, ...)
}
\arguments{
\item{name}{A palette name from the lists below}

\item{n}{Number of different colors in the palette, minimum 3, 
maximum depending on palette}

\item{palette}{string}

\item{reverse}{logical}

\item{...}{Other arguments passed on to \code{\link{continuous_scale}}
to control name, limits, breaks, labels and so forth.}
}
\description{
Wrapper of brewer.pal()

Shortcut of scale_colour_gradientn(colours=brewer.pal(n, name))

Shortcut of scale_fill_gradientn(colours=brewer.pal(n, name))
}
\examples{
ggplot2::scale_colour_gradientn(colours=brewer_palette('RdBu'))
}

