% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{isdir}
\alias{isdir}
\alias{splitext}
\alias{this_file}
\title{A shortcut}
\usage{
isdir(...)

splitext(path)

this_file()
}
\arguments{
\item{...}{character vectors conatining file paths}

\item{path}{a string}
}
\value{
a logical vector

a character vector (root, ext)

a character; NULL in interactive mode
}
\description{
A shortcut

os.path.splitext() in Python

script path in Rscript
}

