% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo.R
\name{node_labels}
\alias{node_labels}
\alias{named_edges}
\alias{ape_layout_unrooted}
\title{Utilities for phylo class of ape package}
\usage{
node_labels(x)

named_edges(x)

ape_layout_unrooted(phy, centering = TRUE, rotate = 0)
}
\arguments{
\item{x}{phylo object}

\item{phy}{phylo object}

\item{centering}{boolean}

\item{rotate}{angle in radian}
}
\description{
Utilities for phylo class of ape package
}
