% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{pipeshell}
\alias{pipeshell}
\alias{\%|\%}
\alias{less}
\alias{grepp}
\alias{\%G\%}
\alias{redirect}
\title{Pipe text to system commands}
\usage{
pipeshell(x, command, ...)

x \%|\% command

less(x, ...)

grepp(x, expr, ...)

x \%G\% expr

redirect(expr, file = "/dev/null")
}
\arguments{
\item{x}{an object to print}

\item{command}{to which the text is sent; string or expression.}

\item{...}{further arguments passed to \code{print}}

\item{expr}{arguments to \code{grep -P}; string or expression.}

\item{file}{a writable \link{connection} or a character string naming the
    file to write to, or \code{NULL} to stop sink-ing.}
}
\description{
Pipe text to system commands
}
\details{
\code{pipeshell} and \code{\%|\%} send \code{x} to \code{command} via shell.

\code{less} print \code{x} in \code{getOption('pager')}. The operation is in-memory and
efficient than \code{page(x, method='print')} that involves a temporary file.

\code{grepp} and \code{\%G\%} are shorthand for \code{x \%|\% "grep -P {expr}"}.

\code{redirect} evaluates \code{expr} while \code{sink(file)} is in operation.
}
\examples{
\dontrun{
mpg \%>\% less()
mpg \%>\% grepp(manual)
mpg \%>\% grepp("manual")
mpg \%>\% pipeshell(wc)

iris \%|\% wc
iris \%|\% "wc -l"
iris \%|\% `grep seto | less`

vir = "seto"
iris \%G\% vir
iris \%G\% !!vir
iris \%G\% `-v seto | less`
}
}
