% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{adjust_print_options}
\alias{adjust_print_options}
\alias{max_print}
\alias{less}
\title{Configure print options.}
\usage{
adjust_print_options(max.print = 30L)

max_print(x, max.print = getOption("max.print"), width = Inf, ...)

less(x, max.print = getOption("max.print"), width = Inf, ...)
}
\arguments{
\item{max.print}{maximum number of rows to print}

\item{x}{an object to print}

\item{width}{maximum number of columns to print}

\item{...}{further arguments passed to \code{print}}
}
\description{
\code{adjust_print_options} sets width and height according to the current environment.

\code{max_print} prints a big tibble as it is.

\code{less} is a shortcut of page(x, method='print').
}
