% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{printdf}
\alias{printdf}
\alias{max_print}
\alias{generate_print_options}
\alias{show_print_options}
\title{Utilities for printing}
\usage{
printdf(
  x,
  n = getOption("pillar.print_max", 30L),
  summarize = getOption("wtl.printdf.summarize", TRUE),
  ...
)

max_print(x, n = getOption("max.print"), width = Inf, ...)

generate_print_options(n = 30L)

show_print_options()
}
\arguments{
\item{x}{an object to print}

\item{n}{maximum number of rows to print}

\item{summarize}{print class(x) and dim(x) if TRUE}

\item{...}{further arguments passed to \code{print}}

\item{width}{maximum number of columns to print}
}
\description{
Utilities for printing
}
\details{
\code{printdf()} is a simple cherry-picking from
\code{tibble:::print.tbl} and \code{data.frame:::print.data.table}.

\code{max_print()} prints as many elements in a big tibble as possible.

\code{generate_print_options()} translates \code{n} and stty size to a list of options.

\code{show_print_options()} prints the current values of print options.
}
\examples{
\dontrun{
printdf(mpg)
max_print(mpg)
}
}
