% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{randint}
\alias{randint}
\title{Shortcut for grid::viewport()}
\usage{
randint(n, a, b)
}
\arguments{
\item{n}{integer}

\item{a}{integer}

\item{b}{integer}
}
\value{
n integers between [a, b]
}
\description{
Shortcut for grid::viewport()
}

