% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{runif.int}
\alias{runif.int}
\title{Generate random integers between a and b}
\usage{
runif.int(n, min = 0L, max = .Machine$integer.max - 1L)
}
\arguments{
\item{n}{number of observations}

\item{min, max}{lower and upper limits of the distribution}
}
\value{
n integer vector
}
\description{
Generate random integers between a and b
}
\examples{
table(runif.int(600L, 1L, 6L))
}
\seealso{
purrr::rdunif
}
