% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_boost_ini}
\alias{read_boost_ini}
\alias{read_pb}
\title{Functions to read files}
\usage{
read_boost_ini(file)

read_pb()
}
\arguments{
\item{file}{filename or text}
}
\description{
\code{read_boost_ini()} reads an INI-like config file of boost::program_options.

\code{read_pb()} reads paste board to a character.
}
