% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{scan_matrix}
\alias{command_args}
\alias{read_boost_ini}
\alias{read_cb}
\alias{scan_matrix}
\title{Scan a simple CSV and make a matrix}
\usage{
scan_matrix(filename, what = double(0), sep = "", skip = 0L,
  header = FALSE)

read_boost_ini(filename)

read_cb(...)

command_args()
}
\arguments{
\item{filename}{a string}

\item{what}{type of data}

\item{sep}{a string}

\item{skip}{an integer}

\item{header}{a logical}

\item{...}{passed to read.table()}
}
\value{
a matrix

a data.frame

a data.frame

a list
}
\description{
Scan a simple CSV and make a matrix

Read an INI-like config file of boost::program_options

read paste board into data.frame

get commandline arguments
}

