% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{rle_df}
\alias{rle_df}
\title{Get rle (run-length encoding) as tibble.}
\usage{
rle_df(x, .name = "value")
}
\arguments{
\item{x}{atomic vector}

\item{.name}{column name in output}
}
\value{
tbl with start and end index columns
}
\description{
Get rle (run-length encoding) as tibble.
}
