% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{split_chr}
\alias{as_chr}
\alias{chr}
\alias{ord}
\alias{rsplit}
\alias{split_chr}
\title{Split a single string into a vector}
\usage{
split_chr(string, pattern = "\\\\s+", n = Inf)

rsplit(string, pattern, n = 42L)

ord(char)

chr(i)

as_chr(...)
}
\arguments{
\item{string}{Input vector. Either a character vector, or something
coercible to one.}

\item{pattern}{Pattern to look for.

  The default interpretation is a regular expression, as described
  in \link[stringi]{stringi-search-regex}. Control options with
  \code{\link{regex}()}.

  Match a fixed string (i.e. by comparing only bytes), using
  \code{\link{fixed}(x)}. This is fast, but approximate. Generally,
  for matching human text, you'll want \code{\link{coll}(x)} which
  respects character matching rules for the specified locale.

  Match character, word, line and sentence boundaries with
  \code{\link{boundary}()}. An empty pattern, "", is equivalent to
  \code{boundary("character")}.}

\item{n}{number of pieces to return.  Default (Inf) uses all
  possible split positions.

  For \code{str_split_fixed}, if n is greater than the number of pieces,
  the result will be padded with empty strings.}

\item{char}{character}

\item{i}{integer}

\item{...}{expressions}
}
\value{
character vectors

a list of character vectors
}
\description{
Split a single string into a vector

Pythonic method to manupilate strings

Pythonic alias of charToRaw()

Pythonic alias of intToUtf8()

Convert expr into character vector
}

