% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{join}
\alias{join}
\alias{split_chr}
\alias{rsplit}
\title{String utility}
\usage{
join(string, sep = "", na.rm = FALSE)

split_chr(string, pattern = "\\\\s+", n = Inf)

rsplit(string, pattern = "\\\\s+", n = 42L)
}
\arguments{
\item{string}{Input vector. Either a character vector, or something
coercible to one.}

\item{sep}{String to insert between input vectors.}

\item{na.rm}{logical}

\item{pattern}{Pattern to look for.

The default interpretation is a regular expression, as described
in \link[stringi:stringi-search-regex]{stringi::stringi-search-regex}. Control options with
\code{\link[stringr:modifiers]{regex()}}.

Match a fixed string (i.e. by comparing only bytes), using
\code{\link[stringr:modifiers]{fixed()}}. This is fast, but approximate. Generally,
for matching human text, you'll want \code{\link[stringr:modifiers]{coll()}} which
respects character matching rules for the specified locale.

Match character, word, line and sentence boundaries with
\code{\link[stringr:modifiers]{boundary()}}. An empty pattern, "", is equivalent to
\code{boundary("character")}.}

\item{n}{number of pieces to return.  Default (Inf) uses all
possible split positions.

For \code{str_split_fixed}, if n is greater than the number of pieces,
the result will be padded with empty strings.}
}
\description{
String utility
}
\details{
\code{join} concatenats strings in a vector.

\code{split_chr} splits a string and return a flattened vector.

\code{rsplit} splits a string from right.
}
