% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{glue}
\alias{glue}
\alias{glue_data}
\alias{str_c_coalesce}
\alias{rsplit}
\alias{as_chr}
\alias{as_code}
\title{String utility}
\usage{
glue(..., .sep = "", .envir = parent.frame())

glue_data(.x, ..., .sep = "", .envir = parent.frame(), .na = "NA")

str_c_coalesce(..., sep = "", collapse = NULL)

rsplit(string, pattern = "\\\\s+", n = 42L)

as_chr(...)

as_code(x)
}
\arguments{
\item{...}{expressions}

\item{.sep, sep}{[\code{character(1)}: \sQuote{""}]\cr Separator used to separate elements.}

\item{.envir}{[\code{environment}: \code{parent.frame()}]\cr Environment to evaluate each expression in. Expressions are
evaluated from left to right. If \code{.x} is an environment, the expressions are
evaluated in that environment and \code{.envir} is ignored. If \code{NULL} is passed, it is equivalent to \code{\link[=emptyenv]{emptyenv()}}.}

\item{.x}{[\code{listish}]\cr An environment, list, or data frame used to lookup values.}

\item{.na}{[\code{character(1)}: \sQuote{NA}]\cr Value to replace \code{NA} values
with. If \code{NULL} missing values are propagated, that is an \code{NA} result will
cause \code{NA} output. Otherwise the value is replaced by the value of \code{.na}.}

\item{collapse}{Optional string used to combine output into single
string. Generally better to use \code{\link[stringr:str_flatten]{str_flatten()}} if you needed this
behaviour.}

\item{string}{Input vector. Either a character vector, or something
coercible to one.}

\item{pattern}{Pattern to look for.

The default interpretation is a regular expression, as described in
\code{vignette("regular-expressions")}. Use \code{\link[stringr:regex]{regex()}} for finer control of the
matching behaviour.

Match a fixed string (i.e. by comparing only bytes), using
\code{\link[stringr:fixed]{fixed()}}. This is fast, but approximate. Generally,
for matching human text, you'll want \code{\link[stringr:coll]{coll()}} which
respects character matching rules for the specified locale.

Match character, word, line and sentence boundaries with
\code{\link[stringr:boundary]{boundary()}}. The empty string, \verb{""``, is equivalent to }boundary("character")`.}

\item{n}{Maximum number of pieces to return. Default (Inf) uses all
possible split positions.

For \code{str_split()}, this determines the maximum length of each element
of the output. For \code{str_split_fixed()}, this determines the number of
columns in the output; if an input is too short, the result will be padded
with \code{""}.}

\item{x}{vector}
}
\description{
\code{glue()} is a shortcut of \code{\link[stringr:str_glue]{stringr::str_glue()}} without trimming.

\code{str_c_coalesce()} tries \code{\link[stringr:str_c]{stringr::str_c()}} first,
and falls back to \code{\link[dplyr:coalesce]{dplyr::coalesce()}}.

\code{rsplit()} splits a string from right.

\code{as_chr()} converts expr into character vector.

\code{as_code()} converts a vector into R code that generates it.
}
