% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{eq_assign_style}
\alias{eq_assign_style}
\alias{styler_style}
\title{Apply tidyverse_style, but use equal sign = for assignment}
\usage{
eq_assign_style(...)

styler_style(
  path = ".",
  ...,
  style = eq_assign_style,
  filetype = "R",
  recursive = TRUE,
  exclude_files = NULL
)
}
\arguments{
\item{...}{Arguments passed on to the \code{style} function,
see \code{\link[styler:tidyverse_style]{tidyverse_style()}} for the default argument.}

\item{path}{Path to a directory with files to transform.}

\item{style}{A function that creates a style guide to use, by default
\code{\link[styler]{tidyverse_style}}. Not used
further except to construct the argument \code{transformers}. See
\code{\link[styler:style_guides]{style_guides()}} for details.}

\item{filetype}{Vector of file extensions indicating which file types should
be styled. Case is ignored, and the \code{.} is optional, e.g. \code{c(".R", ".Rmd")}, or \code{c("r", "rmd")}. Supported values (after standardization) are:
"r", "rprofile", "rmd", "rmarkdown", "rnw", "qmd". Rmarkdown is treated as
Rmd.}

\item{recursive}{A logical value indicating whether or not files in
sub directories of \code{path} should be styled as well.}

\item{exclude_files}{Character vector with paths to files that should be
excluded from styling.}
}
\description{
Apply tidyverse_style, but use equal sign = for assignment
}
