% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_clean}
\alias{theme_clean}
\alias{theme_wtl}
\alias{axis_line}
\title{Incomplete theme that does not overwrite other elements}
\usage{
theme_clean()

theme_wtl(base_size = 12, base_family = "sans")

axis_line(colour = NULL, size = NULL, linetype = NULL, lineend = NULL,
  arrow = NULL, inherit.blank = FALSE)
}
\arguments{
\item{base_size, base_family}{see ggplot2::ggtheme}

\item{colour}{Line/border colour. Color is an alias for colour.}

\item{size}{Line/border size in mm; text size in pts.}

\item{linetype}{Line type. An integer (0:8), a name (blank, solid,
dashed, dotted, dotdash, longdash, twodash), or a string with
an even number (up to eight) of hexadecimal digits which give the
lengths in consecutive positions in the string.}

\item{lineend}{Line end Line end style (round, butt, square)}

\item{arrow}{Arrow specification, as created by \code{\link[grid]{arrow}}}

\item{inherit.blank}{Should this element inherit the existence of an
\code{element_blank} among its parents? If \code{TRUE} the existence of
a blank element among its parents will cause this element to be blank as
well. If \code{FALSE} any blank parent element will be ignored when
calculating final element state.}
}
\description{
Incomplete theme that does not overwrite other elements

Complete theme based on theme_bw()

Set L-shaped axes
}
