% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibbled.R
\name{rle_df}
\alias{rle_df}
\alias{rle_ranges}
\alias{table_df}
\title{Tibbled version of base R functions}
\usage{
rle_df(x)

rle_ranges(...)

table_df(..., name = "n")
}
\arguments{
\item{x, ...}{Atomic vector}

\item{name}{The name of the new column in the output.}
}
\description{
\code{rle_df()} returns a tibbled rle.

\code{rle_ranges()} returns an IRanges-like tibble.

\code{table_df()} returns a tibbled table.
}
\seealso{
\code{\link[=rle]{rle()}}

\code{\link[=table]{table()}}, \code{\link[dplyr:count]{dplyr::count()}}
}
