% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibbled.R
\name{rle_df}
\alias{rle_df}
\alias{table_df}
\title{Tibbled version of base R functions}
\usage{
rle_df(x, name = "value")

table_df(x, name = "n")
}
\arguments{
\item{x}{atomic vector}

\item{name}{column name in output}
}
\description{
Tibbled version of base R functions
}
\seealso{
\code{\link[=rle]{rle()}}

\code{\link[=table]{table()}}, \code{\link[dplyr:count]{dplyr::count()}}
}
