% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{str_today}
\alias{str_today}
\alias{str_now}
\title{Snippet for time in R}
\usage{
str_today(format = "\%Y\%m\%d")

str_now(format = "\%Y\%m\%d_\%H\%M\%S")
}
\arguments{
\item{format}{string}
}
\description{
Snippet for time in R
}
