% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{easierprof}
\alias{easierprof}
\alias{getenv}
\alias{reprex_tidyverse}
\title{Utility functions}
\usage{
easierprof(expr, interval = 0.02, memory = FALSE)

getenv(pattern = NULL)

reprex_tidyverse(venue = "r", n = 8L, html_preview = FALSE)
}
\arguments{
\item{expr}{R expression}

\item{interval}{numeric}

\item{memory}{logical}

\item{pattern}{R expression}

\item{venue}{Character. Must be one of the following (case insensitive):
\itemize{
\item "gh" for \href{https://github.github.com/gfm/}{GitHub-Flavored Markdown}, the
default
\item "r" for a runnable R script, with commented output interleaved. Also useful
for \href{https://slack.com/intl/en-ca/slack-tips/share-code-snippets}{Slack code snippets};
select "R" from the "Type" drop-down menu to enjoy nice syntax
highlighting.
\item "rtf" for
\href{https://en.wikipedia.org/wiki/Rich_Text_Format}{Rich Text Format}
(not supported for un-reprexing)
\item "html" for an HTML fragment suitable for inclusion in a larger HTML
document (not supported for un-reprexing)
\item "slack" for pasting into a Slack message. Optimized for people who opt out
of Slack's WYSIWYG interface. Go to
\strong{Preferences > Advanced > Input options} and select "Format messages with
markup". (If there is demand for a second Slack venue optimized for use
with WYSIWYG, please open an issue to discuss.)
\item "so" for
\href{https://stackoverflow.com/editing-help#syntax-highlighting}{Stack Overflow Markdown}.
Note: this is just an alias for "gh", since Stack Overflow started to
support CommonMark-style fenced code blocks in January 2019.
\item "ds" for Discourse, e.g.,
\href{https://community.rstudio.com}{community.rstudio.com}. Note: this is
currently just an alias for "gh".
}}

\item{n}{passed to \code{pillar.print_max}}

\item{html_preview}{Logical. Whether to show rendered output in a viewer
(RStudio or browser). Always \code{FALSE} in a noninteractive session. Read more
about \code{\link[reprex:opt]{opt()}}.}
}
\description{
\code{easierprof()} is a simple wrapper of \code{\link[utils:Rprof]{utils::Rprof()}} and \code{\link[utils:summaryRprof]{utils::summaryRprof()}}.

\code{getenv()} converts a result of \code{\link[=Sys.getenv]{Sys.getenv()}} to data.frame.

\code{reprex_tidyverse()} wraps \code{\link[reprex:reprex]{reprex::reprex()}}
}
