% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{easierprof}
\alias{easierprof}
\alias{getenv}
\alias{reprex_tidyverse}
\title{Utility functions}
\usage{
easierprof(expr, interval = 0.02, memory = FALSE)

getenv(pattern = NULL)

reprex_tidyverse(n = 8L, venue = "r", show = FALSE)
}
\arguments{
\item{expr}{R expression}

\item{interval}{numeric}

\item{memory}{logical}

\item{pattern}{R expression}

\item{n}{maximum number of rows to print}

\item{venue}{character; see \link[reprex:reprex]{reprex::reprex}}

\item{show}{logical}
}
\description{
Utility functions
}
\details{
\code{easierprof} is a simple wrapper of \code{Rprof()} and \code{summaryRprof()}.

\code{getenv} converts a result of \code{Sys.getenv()} to data.frame.

\code{reprex_tidyverse} wraps \link[reprex:reprex]{reprex::reprex}
}
