% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{easierprof}
\alias{easierprof}
\alias{adjust_print_options}
\alias{max_print}
\alias{less}
\alias{refresh}
\title{Simple wrapper of Rprof() and summaryRprof()}
\usage{
easierprof(expr, interval = 0.02, memory = FALSE)

adjust_print_options(max.print = 30L)

max_print(x, max.print = getOption("max.print"), width = Inf, ...)

less(x, max.print = getOption("max.print"), width = Inf, ...)

refresh(pkg = "rwtl", dir = "~/git")
}
\arguments{
\item{expr}{R expression}

\item{interval}{numeric}

\item{memory}{logical}

\item{max.print}{maximum number of rows to print}

\item{x}{an object used to select a method.}

\item{width}{integer}

\item{...}{further arguments passed to or from other methods.}

\item{pkg}{repository name}

\item{dir}{parent directory of the repository}
}
\value{
summaryRprof()
}
\description{
Simple wrapper of Rprof() and summaryRprof()

Set width and height according to the current environment

Print big tibble as it is

Shortcut of page(x, method='print')

Workaround to achieve devtools::document(..., export_all=FALSE)
}
