% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{easierprof}
\alias{easierprof}
\title{Utility functions}
\usage{
easierprof(expr, interval = 0.02, memory = FALSE)
}
\arguments{
\item{expr}{R expression}

\item{interval}{numeric}

\item{memory}{logical}
}
\description{
\code{easierprof} is a simple wrapper of \code{Rprof()} and \code{summaryRprof()}.
}
