% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{refresh}
\alias{refresh}
\alias{easierprof}
\title{Utility functions}
\usage{
refresh(pkg = "rwtl", dir = "~/git")

easierprof(expr, interval = 0.02, memory = FALSE)
}
\arguments{
\item{pkg}{repository name}

\item{dir}{parent directory of the repository}

\item{expr}{R expression}

\item{interval}{numeric}

\item{memory}{logical}
}
\description{
\code{refresh} is a shortcut for \code{devtools::document(..., export_all=FALSE)}.

\code{easierprof} is a simple wrapper of \code{Rprof()} and \code{summaryRprof()}.
}
