% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{easierprof}
\alias{easierprof}
\alias{getenv}
\title{Utility functions}
\usage{
easierprof(expr, interval = 0.02, memory = FALSE)

getenv(pattern = NULL)
}
\arguments{
\item{expr}{R expression}

\item{interval}{numeric}

\item{memory}{logical}

\item{pattern}{R expression}
}
\description{
Utility functions
}
\details{
\code{easierprof} is a simple wrapper of \code{Rprof()} and \code{summaryRprof()}.

\code{getenv} converts a result of \code{Sys.getenv()} to data.frame.
}
