% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{easierprof}
\alias{adjust_width}
\alias{easierprof}
\alias{less}
\alias{refresh}
\title{Simple wrapper of Rprof() and summaryRprof()}
\usage{
easierprof(expr, interval = 0.02, memory = FALSE)

adjust_width(width = Sys.getenv("COLUMNS"))

less(x, method = c("print", "dput"), max.print = getOption("max.print"),
  ...)

refresh(pkg = "rwtl", dir = "~/git")
}
\arguments{
\item{expr}{R expression}

\item{interval}{numeric}

\item{memory}{logical}

\item{width}{integer}

\item{x}{An \R object, or a character string naming an object.}

\item{method}{The default method is to dump the object \emph{via}
    \code{\link{dput}}.  An alternative is to use \code{print} and
    capture the output to be shown in the pager.  Can be abbreviated.}

\item{max.print}{maximum number of rows to print}

\item{...}{additional arguments for \code{\link{dput}},
    \code{\link{print}} or \code{\link{file.show}} (such as \code{title}).}

\item{pkg}{repository name}

\item{dir}{parent directory of the repository}
}
\value{
summaryRprof()
}
\description{
Simple wrapper of Rprof() and summaryRprof()

Set options(width) according to the current environment

Shortcut of page(x, method='print')

Workaround to achieve devtools::document(..., export_all=FALSE)
}

