% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{easierprof}
\alias{adjust_width}
\alias{easierprof}
\alias{refresh}
\title{Simple wrapper of Rprof() and summaryRprof()}
\usage{
easierprof(expr, interval = 0.02, memory = FALSE)

adjust_width(width = Sys.getenv("COLUMNS"))

refresh(pkg = "rwtl", dir = "~/git")
}
\arguments{
\item{expr}{R expression}

\item{interval}{numeric}

\item{memory}{logical}

\item{width}{integer}

\item{pkg}{repository name}

\item{dir}{parent directory of the repository}
}
\value{
summaryRprof()
}
\description{
Simple wrapper of Rprof() and summaryRprof()

Set options(width) according to the current environment

Workaround to achieve devtools::document(..., export_all=FALSE)
}

