% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{twist}
\alias{twist}
\alias{split_consecutive}
\title{Utilities for vector}
\usage{
twist(x, n)

split_consecutive(x)
}
\arguments{
\item{x}{a vector}

\item{n}{an integer}
}
\description{
Utilities for vector
}
\details{
\code{\link[=twist]{twist()}} cycles a vector

\code{\link[=split_consecutive]{split_consecutive()}} groups consecutive integers.
}
\seealso{
\code{\link[dplyr:lead]{dplyr::lead()}}, \code{\link[dplyr:lag]{dplyr::lag()}}
}
