% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{twist}
\alias{twist}
\alias{split_consecutive}
\title{Utilities for vector}
\usage{
twist(x, n)

split_consecutive(x)
}
\arguments{
\item{x}{a vector}

\item{n}{an integer}
}
\description{
\code{twist()} cycles a vector.

\code{split_consecutive()} groups consecutive integers.
}
\seealso{
\code{\link[dplyr:lead-lag]{dplyr::lead()}}, \code{\link[dplyr:lead-lag]{dplyr::lag()}}
}
